refactor: move all bytea wrapper into bytea.rs
