+++ src/servers/src/postgres/types.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod bytea;
 mod interval;
 
 use std::collections::HashMap;
